Add automatic processing for all RPC failures
This makes all RPC failures start by passing through Game.Failure, and
giving it a change to handle them.  If it doesn't, the specified failure
callback will be called.  If that callback doesn't handle the error,
display a generic error message.

In most cases, this removes the need to have a failure callback at all.