Bold from *'s in messages only if
a) the first * is not followed by a space
b) there are no \n's in between the *'s

For example:

* not bold
* still not bold *
*this is bold*
*this is bold, too *
* not bold *but this is*

This allows bullets and some use of *'s in things like mathematical equations.
It does, however, mean that if you want to bold multiple paragraphs, you need
to put *'s around each paragraph individually.  This is not likely a real
issue.