Change default number of ships to send in fleet tab to 1.
No one sends zero ships, so the default is wrong 100% of the time.  This
will be useful at least part of the time.

Maybe better would be to figure out what the maximum is based on the type
of ship (e.g., supply pods and colony ships can only send one at a time,
but sweepers and snarks have a max of however many are available, or 600,
whichever is smaller), but this would require much more work for little
extra payoff.

Due to auto-grouping, accidentally hitting send with only one ship when
more is intended isn't a big deal, just send more, they'll group together.